LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-tojson-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


FAIL toJSON Can't find variable: trustedTypes
